requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
ibm-cloud-sdk-core==0.5.1
idna==2.8
pbr==2.1.0
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.21.0
urllib3==1.25.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
websocket-client==0.48.0
